# src/main/resources/openapi/yaml/api-order-interbank.yaml
---
openapi: 3.1.0
info:
  version: '1'
  title: API BS Order Interbank
  description: API diseñada para el reto Técnico con Interbank
servers:
  - url: /http://localhost:8080/
    description: SwaggerHub API Auto Mocking
tags:
  - name: order
    description: Todo sobre los Pedidos del Sistema
  - name: product
    description: Todo sobre los Productos del Sistema.
paths:
  /orders:
    post:
      tags:
        - order
      summary: Create a Order
      description: Add a new Order
      operationId: registerOrder
      requestBody:
        description: Create a new Order in the System INTERBANK
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderSend'
          application/xml:
            schema:
              $ref: '#/components/schemas/OrderSend'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OrderSend'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItem'
            application/xml:
              schema:
                $ref: '#/components/schemas/OrderItem'
        "400":
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "404":
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "409":
          description: Conflicto de negocio (stock insuficiente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
  /products:
    get:
      tags:
        - product
      summary: List Products.
      description: Return list products
      operationId: getProducts
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
        "400":
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "404":
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
    post:
      tags:
        - product
      summary: Create a Product
      description: Add a new Product
      operationId: registerProduct
      requestBody:
        description: Create a new product in the System
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductRequest'
        required: true
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        "400":
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "404":
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "409":
          description: Conflicto de negocio (stock insuficiente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
  /products/{name}:
    get:
      tags:
        - product
      summary: Find by product for Name
      description: Return product for Name of database
      operationId: getProductByName
      parameters:
        - name: name
          in: path
          description: name product of database
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        "400":
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "404":
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
    put:
      tags:
        - product
      summary: Update product
      description: Update product in the System.
      operationId: updateProduct
      parameters:
        - name: name
          in: path
          description: name product of database
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent product in the system.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductRequest'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        "400":
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "404":
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "409":
          description: Conflicto de negocio (stock insuficiente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
    delete:
      tags:
        - product
      summary: Delete a product.
      description: delete a product.
      operationId: deleteProduct
      parameters:
        - name: name
          in: path
          description: Product name to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        "400":
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "401":
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "404":
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "409":
          description: Conflicto de negocio (stock insuficiente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
components:
  schemas:
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID del Producto.
          format: int64
          examples: [1]
        name:
          type: string
          description: Nombre del Producto.
          maxLength: 50
          pattern: '^[A-Za-z]+$'
          examples: ["Inka Kola"]
        price:
          type: number
          description: Precio del Producto.
          format: double
          examples: [2.50]
        stock:
          type: integer
          description: Stock del Producto.
          format: int64
          examples: [6]
    ProductRequest:
      type: object
      properties:
        name:
          type: string
          description: Nombre del Producto.
          maxLength: 50
          pattern: '^[A-Za-z]+$'
          examples: ["Inka Kola"]
        price:
          type: number
          description: Precio del Producto.
          format: double
          examples: [2.50]
        stock:
          type: integer
          description: Stock del Producto.
          format: int64
          examples: [6]
    OrderRequest:
      type: object
      properties:
        listOrder:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ProductItem'
    OrderSend:
      type: object
      properties:
        date:
          type: string
          description: Fecha realizada del Pedido.
          pattern: '^\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])$'
          examples: ["2025-12-26"]
        products:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ProductSend'
    ProductSend:
      type: object
      properties:
        name:
          type: string
          description: Nombre del Producto.
          maxLength: 50
          pattern: '^[A-Za-z]+$'
          examples: ["Inka Kola"]
        price:
          type: number
          description: Precio del Producto.
          format: double
          examples: [2.50]
        lot:
          type: integer
          description: Cantidad del Producto.
          format: int64
          examples: [1]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          description: ID del Pedido.
          format: int64
          examples: [1]
        date:
          type: string
          description: Fecha realizada del Pedido.
          pattern: '^\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])$'
          examples: ["2025-12-26"]
        total:
          type: number
          description: Total del Pedido.
          format: double
          examples: [1020.50]
        state:
          type: string
          description: Estado del Pedido.
          enum:
            - CREATED
            - CONFIRMED
            - CANCELLED
          examples: [CONFIRMED]
    ProductItem:
      type: object
      properties:
        id:
          type: integer
          description: ID del Producto.
          format: int64
          examples: [1]
        name:
          type: string
          description: Nombre del Producto.
          maxLength: 50
          pattern: '^[A-Za-z]+$'
          examples: ["Inka Kola"]
        price:
          type: number
          description: Precio del Producto.
          format: double
          examples: [2.50]
        stock:
          type: integer
          description: Stock del Producto.
          format: int64
          examples: [6]
    ApiException:
      title: ApiException
      type: object
      required:
        - code
        - description
      properties:
        code:
          type: string
          description: Codigo de error de Sistema.
          examples: [XXX]
        description:
          type: string
          description: Descripción del error de Sistema.
          examples: [Error al llamar al servicio]
        errorType:
          type: string
          description: Tipo de Error de Sistema.
          examples: [TECHNICAL]
        exceptionDetails:
          type: array
          description: Lista de detalles del error.
          items:
            $ref: '#/components/schemas/ApiExceptionDetail'
        properties:
          type: object
          properties: {}
          description: Lista de propiedades especiales.
      description: Datos de error del Sistema.
    ApiExceptionDetail:
      title: ApiExceptionDetail
      type: object
      properties:
        code:
          type: string
          description: Codigo de error del Detalle.
          examples: [MXXXX01]
        component:
          type: string
          description: Nombre del componente de falla.
          examples: [api-ms-order]
        description:
          type: string
          description: Descripción del Detalle.
          examples: [Codigo invalido]
      description: Datos del error técnico.